Refactor : response format 패키지 정리
