+++ src/main/java/com/kusitms/forpet/controller/CertifyController.java
@@ -2,7 +2,7 @@ package com.kusitms.forpet.controller;
 
 import com.kusitms.forpet.domain.PetCard;
 import com.kusitms.forpet.domain.User;
-import com.kusitms.forpet.dto.ApiResponse;
+import com.kusitms.forpet.dto.response.ApiResponse;
 import com.kusitms.forpet.dto.SignUpDto;
 import com.kusitms.forpet.security.TokenProvider;
 import com.kusitms.forpet.service.PetCardService;
@@ -13,7 +13,6 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.util.HashMap;
 import java.util.Map;
