--- src/main/java/com/kusitms/forpet/dto/ApiResponse.java
@@ -1,67 +1,0 @@
-package com.kusitms.forpet.dto;
-
-import lombok.*;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Data
-public class ApiResponse<T> {
-    private final static int SUCCESS = 200;
-    public static final int CREATED = 201;
-    private final static int NOT_FOUND = 404;
-    private final static int FAILED = 500;
-    private final static int BAD_REQUEST = 400;
-    private final static String SUCCESS_MESSAGE = "SUCCESS";
-    private final static String CREATED_MESSAGE = "CREATED";
-    private final static String UPDATED_MESSAGE = "UPDATED";
-    private final static String NOT_FOUND_MESSAGE = "NOT FOUND";
-    private final static String FAILED_MESSAGE = "서버에서 오류가 발생하였습니다.";
-    private final static String INVALID_ACCESS_TOKEN = "Invalid access token.";
-    private final static String INVALID_REFRESH_TOKEN = "Invalid refresh token.";
-    private final static String NOT_EXPIRED_TOKEN_YET = "Not expired token yet.";
-    private final static String BAD_REQUEST_MESSAGE = "Bad Request";
-
-    private final ApiResponseHeader header;
-    private final Map<String, T> body;
-
-    public static <T> ApiResponse<T> success(String name, T body) {
-        Map<String, T> map = new HashMap<>();
-        map.put(name, body);
-
-        return new ApiResponse(new ApiResponseHeader(SUCCESS, SUCCESS_MESSAGE), map);
-    }
-
-    public static <T> ApiResponse<T> created(String name, T body) {
-        Map<String, T> map = new HashMap<>();
-        map.put(name, body);
-
-        return new ApiResponse(new ApiResponseHeader(CREATED, CREATED_MESSAGE), map);
-    }
-    public static <T> ApiResponse<T> updated(String name, T body) {
-        Map<String, T> map = new HashMap<>();
-        map.put(name, body);
-
-        return new ApiResponse(new ApiResponseHeader(CREATED, UPDATED_MESSAGE), map);
-    }
-    public static <T> ApiResponse<T> fail() {
-        return new ApiResponse(new ApiResponseHeader(FAILED, FAILED_MESSAGE), null);
-    }
-
-    public static <T> ApiResponse<T> invalidAccessToken() {
-        return new ApiResponse(new ApiResponseHeader(FAILED, INVALID_ACCESS_TOKEN), null);
-    }
-
-    public static <T> ApiResponse<T> invalidRefreshToken() {
-        return new ApiResponse(new ApiResponseHeader(FAILED, INVALID_REFRESH_TOKEN), null);
-    }
-
-    public static <T> ApiResponse<T> notExpiredTokenYet() {
-        return new ApiResponse(new ApiResponseHeader(FAILED, NOT_EXPIRED_TOKEN_YET), null);
-    }
-
-
-    public static <T> ApiResponse<T> badRequest() {
-        return new ApiResponse<>(new ApiResponseHeader(BAD_REQUEST, BAD_REQUEST_MESSAGE), null);
-    }
-}
