+++ src/main/java/com/kusitms/forpet/dto/response/ApiResponseHeader.java
@@ -1,4 +1,4 @@
-package com.kusitms.forpet.dto;
+package com.kusitms.forpet.dto.response;
 
 import lombok.*;
